refactor: 💡 Split out StatusIndicator into it's own Component
The prevouis style is not a great pattern and would cause
StatusIndicator to remount on state change of Lozenge. This is not a
problem for now but could be if StatusIndicator had it's own state.
